+++ examples/robot_control_loop.py
@@ -64,12 +64,12 @@ def run_cartesian_mode():
     dt = 1.0 / ctrl_freq
 
     # Randomly initialize robot state (actual pose from sensors/encoders)
-    state_pos_xyz = np.array([0.3, 0.0, 0.5], dtype=np.float32)    # Position in meters (x, y, z)
-    state_quat_wxyz = np.array([1.0, 0.0, 0.0, 0.0], dtype=np.float32)  # Quaternion (w, x, y, z)
+    state_pos_xyz = np.array([0.3, 0.0, 0.5], dtype=np.float64)    # Position in meters (x, y, z)
+    state_quat_wxyz = np.array([1.0, 0.0, 0.0, 0.0], dtype=np.float64)  # Quaternion (w, x, y, z)
 
     # Randomly initialize target pose (commanded poses)
-    target_pos_xyz = np.array([0.35, 0.05, 0.55], dtype=np.float32)
-    target_quat_wxyz = np.array([1.0, 0.0, 0.0, 0.0], dtype=np.float32)
+    target_pos_xyz = np.array([0.35, 0.05, 0.55], dtype=np.float64)
+    target_quat_wxyz = np.array([1.0, 0.0, 0.0, 0.0], dtype=np.float64)
 
     print(f"[CARTESIAN MODE] Robot control loop running at {ctrl_freq} Hz. Press Ctrl+C to stop.")
     try:
@@ -78,12 +78,12 @@ def run_cartesian_mode():
 
             # Get actual robot state (replace with real robot API)
             # e.g., state_pos_xyz, state_quat_wxyz = robot.get_eef_pose()
-            state_pos_xyz += np.random.uniform(-0.001, 0.001, 3).astype(np.float32)
+            state_pos_xyz += np.random.uniform(-0.001, 0.001, 3).astype(np.float64)
             state_timestamp = time.monotonic()
 
             # Get commanded target (replace with controller output)
             # e.g., target_pos_xyz, target_quat_wxyz = controller.get_target()
-            target_pos_xyz += np.random.uniform(-0.002, 0.002, 3).astype(np.float32)
+            target_pos_xyz += np.random.uniform(-0.002, 0.002, 3).astype(np.float64)
             target_timestamp = time.monotonic()
 
             # Check if main process has requested recording
@@ -91,14 +91,14 @@ def run_cartesian_mode():
                 # Log actual state (measured from robot)
                 logger.log_state(
                     state_timestamp=state_timestamp,   # When state was measured
-                    state_pos_xyz=state_pos_xyz,       # Actual position (3,) float32
-                    state_quat_wxyz=state_quat_wxyz    # Actual orientation (4,) float32, w-first
+                    state_pos_xyz=state_pos_xyz,       # Actual position (3,) float64
+                    state_quat_wxyz=state_quat_wxyz    # Actual orientation (4,) float64, w-first
                 )
                 # Log commanded target (from controller)
                 logger.log_target(
                     target_timestamp=target_timestamp, # When target was commanded
-                    target_pos_xyz=target_pos_xyz,     # Target position (3,) float32
-                    target_quat_wxyz=target_quat_wxyz  # Target orientation (4,) float32, w-first
+                    target_pos_xyz=target_pos_xyz,     # Target position (3,) float64
+                    target_quat_wxyz=target_quat_wxyz  # Target orientation (4,) float64, w-first
                 )
 
             elapsed = time.monotonic() - loop_start
@@ -141,12 +141,12 @@ def run_joint_mode():
     dt = 1.0 / ctrl_freq
 
     # Initialize robot state (actual measurements from sensors)
-    state_joint_pos = np.array([0.0, -0.785, 0.0, -2.356, 0.0, 1.571, 0.785], dtype=np.float32)  # 7 joints
-    state_pos_xyz = np.array([0.3, 0.0, 0.5], dtype=np.float32)       # Inferred from FK
-    state_quat_wxyz = np.array([1.0, 0.0, 0.0, 0.0], dtype=np.float32)
+    state_joint_pos = np.array([0.0, -0.785, 0.0, -2.356, 0.0, 1.571, 0.785], dtype=np.float64)  # 7 joints
+    state_pos_xyz = np.array([0.3, 0.0, 0.5], dtype=np.float64)       # Inferred from FK
+    state_quat_wxyz = np.array([1.0, 0.0, 0.0, 0.0], dtype=np.float64)
 
     # Initialize target joint positions (commanded by controller)
-    target_joint_pos = np.array([0.1, -0.8, 0.05, -2.4, 0.0, 1.6, 0.8], dtype=np.float32)
+    target_joint_pos = np.array([0.1, -0.8, 0.05, -2.4, 0.0, 1.6, 0.8], dtype=np.float64)
 
     print(f"[JOINT MODE] Joint-controlled arm loop running at {ctrl_freq} Hz. Press Ctrl+C to stop.")
     print("This example logs both joint positions (control) and EEF pose (inferred via FK)")
@@ -157,16 +157,16 @@ def run_joint_mode():
 
             # Get actual robot state from sensors
             # In real robot: state_joint_pos = robot.get_joint_pos()
-            state_joint_pos += np.random.uniform(-0.001, 0.001, 7).astype(np.float32)
+            state_joint_pos += np.random.uniform(-0.001, 0.001, 7).astype(np.float64)
 
             # Compute EEF pose via forward kinematics
             # In real robot: state_pos_xyz, state_quat_wxyz = robot.forward_kinematics(state_joint_pos)
-            state_pos_xyz += np.random.uniform(-0.0005, 0.0005, 3).astype(np.float32)
+            state_pos_xyz += np.random.uniform(-0.0005, 0.0005, 3).astype(np.float64)
             state_timestamp = time.monotonic()
 
             # Get commanded target joint positions from controller
             # In real robot: target_joint_pos = controller.get_target_joints()
-            target_joint_pos += np.random.uniform(-0.002, 0.002, 7).astype(np.float32)
+            target_joint_pos += np.random.uniform(-0.002, 0.002, 7).astype(np.float64)
             target_timestamp = time.monotonic()
 
             # Check if main process has requested recording
@@ -174,15 +174,15 @@ def run_joint_mode():
                 # Log actual state (measured joints + inferred EEF pose)
                 logger.log_state(
                     state_timestamp=state_timestamp,
-                    state_joint_pos=state_joint_pos,       # Actual joint positions (7,) float32
-                    state_pos_xyz=state_pos_xyz,           # EEF position from FK (3,) float32
-                    state_quat_wxyz=state_quat_wxyz        # EEF orientation from FK (4,) float32
+                    state_joint_pos=state_joint_pos,       # Actual joint positions (7,) float64
+                    state_pos_xyz=state_pos_xyz,           # EEF position from FK (3,) float64
+                    state_quat_wxyz=state_quat_wxyz        # EEF orientation from FK (4,) float64
                 )
 
                 # Log commanded target (joint positions only, since that's what we control)
                 logger.log_target(
                     target_timestamp=target_timestamp,
-                    target_joint_pos=target_joint_pos      # Target joint positions (7,) float32
+                    target_joint_pos=target_joint_pos      # Target joint positions (7,) float64
                 )
 
             elapsed = time.monotonic() - loop_start
